requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
docx2txt
python-dotenv
langchain-groq
fastapi
uvicorn
streamlit
python-multipart
slowapi
